completed front end store for delete videos
